0501 Order API

+++ Backend/mydaydeli/src/main/java/com/myday/model/Order.java
@@ -41,5 +41,6 @@ public class Order {
 //    private Payment payment;
 
     private int totalItem;
-    private int totalPrice;
+
+    private Long totalPrice;
 }
